requirements.txt
aiogram~=2.25.1
pydantic~=1.10.6
python-dotenv~=1.0.0
asyncpg~=0.27.0